feat(multi-org): remove per-call orgId for proxied datasource tools
Upstream #1001 (released in v1.1.0) replaced the per-session proxied state this
built on with a shared, credential-keyed proxiedToolSet owned by the
ToolManager, reference counted and tracking in-flight calls. The org-keyed
client map, defaultOrgID, and cross-org discovery added here edit structures
that no longer exist, so they cannot be merged forward as-is -- they need
reworking against that design. Removing them first keeps merging v1.1.0
tractable; the feature can be reworked and restored separately.

Restores proxied_client.go, proxied_handler.go, proxied_tools.go, session.go,
session_test.go and session_reaper_test.go to their pre-multi-org state, and
drops the proxied multi-org tests and the org-2 Tempo seed fixture.

The org-enumeration helpers move to orgid.go rather than disappearing, since
user_info depends on them: OrgInfo, ListUserOrgs, UserInfo, CurrentUserInfo and
grafanaGetJSON. They sit beside the orgId argument plumbing because
ListUserOrgs reports exactly the set of values that argument can name.
fetchUserOrgIDs, resolveDefaultOrgID and DefaultOrgID are deleted -- cross-org
discovery was their only caller, and DefaultOrgID had none at all.

Docs and the flag help now say "every native tool" rather than "every tool":
proxied tools register via server.AddTool directly, bypassing resolveTool, so
they never receive the injected orgId argument.
