+++ proxied_client.go
@@ -24,7 +24,6 @@ type ProxiedClient struct {
 	DatasourceUID  string
 	DatasourceName string
 	DatasourceType string
-	OrgID          int64 // Grafana org this client is scoped to
 	Client         *mcp_client.Client
 	Tools          []mcp.Tool
 	mutex          sync.RWMutex
@@ -42,13 +41,8 @@ func contextCauseOrErr(ctx context.Context, err error) error {
 }
 
 // NewProxiedClient creates a new connection to a remote MCP server
-func NewProxiedClient(ctx context.Context, orgID int64, datasourceUID, datasourceName, datasourceType, mcpEndpoint string) (*ProxiedClient, error) {
+func NewProxiedClient(ctx context.Context, datasourceUID, datasourceName, datasourceType, mcpEndpoint string) (*ProxiedClient, error) {
 	config := GrafanaConfigFromContext(ctx)
-	// Bake the org into this client's transport so every request to the
-	// datasource proxy carries X-Grafana-Org-Id: orgID, independent of the
-	// per-call context.
-	config.OrgID = orgID
-	ctx = WithGrafanaConfig(ctx, config)
 	logger := config.LoggerOrDefault()
 
 	initCtx, cancel := context.WithTimeoutCause(ctx, mcpClientInitTimeout,
@@ -103,7 +97,6 @@ func NewProxiedClient(ctx context.Context, orgID int64, datasourceUID, datasourc
 		DatasourceUID:  datasourceUID,
 		DatasourceName: datasourceName,
 		DatasourceType: datasourceType,
-		OrgID:          orgID,
 		Client:         mcpClient,
 		Tools:          toolsResult.Tools,
 	}, nil
