+++ testdata/orgs-seed.sh
@@ -39,21 +39,3 @@ seed_instance() {
 
 seed_instance "${GRAFANA_URL:-http://grafana:3000}" "mcp-orgid-ns" "OrgID NS Dashboard"
 seed_instance "${LEGACY_GRAFANA_URL:-http://grafana-legacy:3000}" "mcp-orgid-legacy" "OrgID Legacy Dashboard"
-
-# Seed a Tempo datasource into the secondary org on the modern instance, so the
-# proxied-tools multi-org integration test can discover an MCP datasource that
-# only exists in a non-default org and route to it via orgId. It points at the
-# same Tempo backend (which serves MCP) as the default org's Tempo.
-seed_org2_tempo() {
-  base="$1"
-  org_id=$(curl -sf -u admin:admin "$base/api/orgs/name/$ORG_NAME" | sed -n 's/.*"id":\([0-9]*\).*/\1/p')
-  if [ -z "$org_id" ]; then
-    return 0
-  fi
-  curl -sf -u admin:admin -H "X-Grafana-Org-Id: $org_id" \
-    -X POST "$base/api/datasources" -H "Content-Type: application/json" \
-    -d '{"name":"Tempo Org2","uid":"tempo-org2","type":"tempo","access":"proxy","url":"http://tempo:3200"}' >/dev/null 2>&1 || true
-  echo "seeded tempo datasource (tempo-org2) in org $ORG_NAME (id=$org_id) on $base"
-}
-
-seed_org2_tempo "${GRAFANA_URL:-http://grafana:3000}"
